requirements.txt
pandas
PyYAML
backtesting
sambo
pytest
#pandas-ta
streamlit
plotly
numpy
quantstats>=0.0.69
scikit-learn>=1.0.0
IPython
google-genai